Major feature expansion across TUI widgets, rendering, and demo showcase
Doctor FrankenTUI diagnostics:
- Massively expand capture.rs (+430 lines) with comprehensive terminal
  capability detection, rendering diagnostics, and environment probing
- Update analytics profiles

Core engine improvements:
- Refine cursor movement, event handling, and input parser
- Overhaul S3-FIFO cache eviction algorithm
- Enhance terminal session management with new PTY handling
- Add PTY terminal session test and reproduction test cases

Demo showcase screen updates:
- Update all demo screens: accessibility_panel, dashboard, determinism_lab,
  explainability_cockpit, file_browser, intrinsic_sizing, layout_inspector,
  log_search, markdown_live_editor, markdown_rich_text, table_theme_gallery,
  terminal_capabilities, visual_effects, widget_gallery
- Refine chrome and theme modules

Extras and visual effects:
- Enhance markdown renderer and syntax highlighting
- Expand visual effects system with doom_melt and quake_console effects
- Add markdown table snapshot tests and markdown image alt text test
- Refresh markdown table snapshots

Layout engine:
- Update grid layout and pane management algorithms
- Add layout coherence bug reproduction test

Text and rendering:
- Refine text markup, rope data structure, text handling, and word wrap
- Add wrap reproduction test
- Update PTY handling, ANSI rendering, buffer management, cell handling,
  diff algorithm, link registry, and presenter

Widget expansions:
- Enhance drag, input, list, log_viewer, paragraph, scrollbar, table,
  textarea, tree, and virtualized widgets
- Add new list_helper.rs utility module
- Add table sort and textarea wide scroll tests
- Update textarea soft wrap test

Runtime improvements:
- Refine program lifecycle, render thread, telemetry, and terminal writer
- Update WASM showcase Cargo.toml and style/theme module

Documentation and tracking:
- Update master-todo-bead-map.md and add definition_of_done.md
- Update UPGRADE_LOG.md and beads issue tracker
- Add reproduction test directories

+++ crates/ftui-core/src/cursor.rs
@@ -207,7 +207,7 @@ impl CursorManager {
             CursorSaveStrategy::Emulated => {
                 if let Some((col, row)) = self.saved_position {
                     // CUP uses 1-indexed coordinates
-                    write!(writer, "\x1b[{};{}H", row + 1, col + 1)
+                    write!(writer, "\x1b[{};{}H", (row as u32) + 1, (col as u32) + 1)
                 } else {
                     Ok(())
                 }
@@ -252,7 +252,7 @@ impl Default for CursorManager {
 /// Returns an error if writing to the output fails.
 pub fn move_to<W: Write>(writer: &mut W, col: u16, row: u16) -> io::Result<()> {
     // CUP uses 1-indexed coordinates
-    write!(writer, "\x1b[{};{}H", row + 1, col + 1)
+    write!(writer, "\x1b[{};{}H", (row as u32) + 1, (col as u32) + 1)
 }
 
 /// Hide the cursor.
